Implemented Dynamic Text Area
